Add first test
