+++ test/Recipe-test.js
@@ -1,0 +1,51 @@
+const chai = require('chai');
+
+const expect = chai.expect;
+var Recipe = require('../src/Recipe')
+
+describe('Recipe', function () {
+            let recipe;
+            beforeEach(function () {
+                var recipeData = {
+                    'id': 595736,
+                    'image': 'test-src',
+                    'ingredients': [{
+                            'id': 20081,
+                            'quantity': {
+                                'amount': 1.5,
+                                'unit': 'c'
+                            }
+                        },
+                        {
+                            'id': 18372,
+                            'quantity': {
+                                'amount': 0.5,
+                                'unit': 'tsp'
+                            }
+                        }
+                    ],
+                    'instructions': [{
+                            'instruction': 'In a large mixing bowl, whisk together the dry ingredients (flour, pudding mix, soda and salt). Set aside.In a large mixing bowl of a stand mixer, cream butter for 30 seconds. Gradually add granulated sugar and brown sugar and cream until light and fluffy.',
+                            'number': 1
+                        },
+                        {
+                            'instruction': 'Add egg and vanilla and mix until combined.',
+                            'number': 2
+                        }
+                    ],
+                    'name': 'Loaded Chocolate Chip Pudding Cookie Cups',
+                    'tags': [
+                        'antipasti',
+                        'starter',
+                        'snack',
+                        'appetizer',
+                        'antipasto',
+                        'hor d\'oeuvre'
+                    ]
+                };
+                recipe = new Recipe(recipeData);
+            });
+
+            it('should be a function', function () {
+                expect(Recipe).to.be.a('function');
+            });
